Update project to exclusively use Typescript
There are a few big changes here to note.  I wasn't able to get the
project running in its current state, so I decided to go fully into the
Typescript conversion of the project, which had the great side-effect of
resolving the issues I was seeing.
All of the components in `server` are now Typescript, and all of the
imports in the examples have been fixed up to suit the new import model.
NextJS and Typescript still have some issues with compatibility.  For
example, I had to create a separate `tsconfig.server.json` file because
Next will overwrite the tsconfig file with parameters that break the
project, as recommended in https://github.com/zeit/next.js/issues/7361
and most of the configuration for a custom server was based on their
example at
https://github.com/zeit/next.js/tree/canary/examples/custom-server-typescript.
All of the testing was manual, for the dev server and the prod build.
Let me know if you have any questions!